test(playback): add targeted tests to kill 6 surviving mutants
Mutation testing on PR #13 found 6 surviving mutants in the new
playback/mod.rs code:

  MISSED  play_state_to_ws -> Default::default()
  MISSED  PlaybackEngine::apply_event -> ()
  MISSED  apply_event: != -> ==  (state-change broadcast guard)
  MISSED  maybe_broadcast_position_update: > -> ==
  MISSED  maybe_broadcast_position_update: > -> <
  MISSED  maybe_broadcast_position_update: > -> >=

New tests that kill each:

- play_state_to_ws_maps_all_variants: exhaustive check of the mapping
  helper across every PlayState variant.
- maybe_broadcast_position_update_boundary_checks: directly sets
  last_now_playing_broadcast to an Instant offset by 499/500/501 ms
  and asserts broadcast / no-broadcast at each boundary. Kills all
  three >= mutants because it pins the behaviour at exactly the
  comparison point.
- apply_event_triggers_state_change_and_broadcast: drives the engine
  through Idle -> WaitingForScene via apply_event directly and asserts
  both the internal state change and the PlaybackStateChanged
  broadcast. Kills the whole-function -> () mutant.
- apply_event_no_broadcast_when_state_unchanged: applies
  VideosAvailable twice in a row; second application must NOT produce
  a broadcast because the state is already WaitingForScene. Kills
  the != -> == guard mutant.

The added tests pushed mod.rs over the 1000-line file-size limit, so
the test module is extracted to a sibling tests.rs file included via
#[cfg(test)] #[path = "tests.rs"] mod tests;. Same pattern used for
routes.rs/routes_tests.rs earlier in this PR.
